chore: move test functions to separated files/packages
